Add reader fuzzer and harden reader against corrupt Parquet files
Introduces PlankReaderFuzzTarget and Plank.Fuzzing.Reader.Target to fuzz
the reader with arbitrary bytes, plus CorruptParquetException for corrupt
file detection. Fixes all 86 crashes found by AFL:

- Propagate unsigned types (ulong/uint) through offsets, sizes, and value
  counts across IParquetReadSource, PageHeader, InternalColumnChunkMetadata,
  InternalRowGroupMetadata, RowGroupToken, and decode methods; cast to int
  only at system API boundaries
- Add ReadI32AsU32/ReadI64AsU64(max) with Remaining to CompactProtocolReader
  to validate footer and page header fields at parse time
- Validate CompressedPageSize against outer buffer after header is consumed
- Add ValidatePlainPayload to guard all fixed-width decode branches
- Replace InvalidDataException with CorruptParquetException throughout
- Fix RowGroupReader to size RowGroupReadContext by schema column count;
  remove upfront column count check; validate lazily in EnumeratePages
- Fix MemoryReadSource/FileReadSource to throw CorruptParquetException
  instead of ArgumentOutOfRangeException for out-of-bounds reads

+++ Plank.Tests/Writer/ColumnStatisticsTests.cs
@@ -450,11 +450,11 @@ internal sealed class ColumnStatisticsTests
             throw new InvalidOperationException("Column index was written before the column chunk.");
         if (column.OffsetIndexOffset <= column.ColumnIndexOffset)
             throw new InvalidOperationException("Offset index was not written after the column index.");
-        var actualStatisticsCount = ReadColumnIndexPageCount(fileBytes, column.ColumnIndexOffset, column.ColumnIndexLength);
+        var actualStatisticsCount = ReadColumnIndexPageCount(fileBytes, (long)column.ColumnIndexOffset, (int)column.ColumnIndexLength);
         if (actualStatisticsCount != expectedPageCount)
             throw new InvalidOperationException(
                 $"Column index page count mismatch. Expected {expectedPageCount}, got {actualStatisticsCount}.");
-        var actualPageCount = ReadOffsetIndexPageCount(fileBytes, column.OffsetIndexOffset, column.OffsetIndexLength);
+        var actualPageCount = ReadOffsetIndexPageCount(fileBytes, (long)column.OffsetIndexOffset, (int)column.OffsetIndexLength);
         if (actualPageCount != expectedPageCount)
             throw new InvalidOperationException(
                 $"Offset index page count mismatch. Expected {expectedPageCount}, got {actualPageCount}.");
@@ -465,8 +465,8 @@ internal sealed class ColumnStatisticsTests
         var bytes = File.ReadAllBytes(path);
         var trailer = bytes.AsSpan(bytes.Length - 8, 8);
         var footerLength = BinaryPrimitives.ReadInt32LittleEndian(trailer[..4]);
-        var footerOffset = bytes.Length - 8 - footerLength;
-        return ParquetMetadataThriftReader.Read(bytes.AsSpan(footerOffset, footerLength), footerOffset);
+        var footerOffset = (ulong)(bytes.Length - 8 - footerLength);
+        return ParquetMetadataThriftReader.Read(bytes.AsSpan((int)footerOffset, footerLength), footerOffset);
     }
 
     static int ReadColumnIndexPageCount(byte[] fileBytes, long offset, int length)
